# mydocs/swagger.yaml
openapi: "3.0.0"
info:
  title: Learn express and Swagger
  description: Course segment about writing docs
  version: 1.1.0
  contact:
    email: fy@gmail.com
    url: "hhtps://fy.dev"

# servers:
#   - url: "https://localhost:4000/api/v1"
#     description: for localhost - secure
#   - url: "http://localhost:4000/api/v1"
#     description: for localhost - regular

servers:
  - url: "{protocol}://localhost:4000/api/{version}"
    description: for localhost
    variables:
      version:
        enum:
          - v1
          - v2
        default: v1
      protocol:
        enum:
          - http
          - https
        default: http

components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: token
    BearerAuth:
      type: http
      scheme: bearer

paths:
  /docs:
    get:
      tags:
        - String
      summary: returns a greet message
      responses: 
        200:
          description: All good and success
          content:
            application/json:
              schema:
                type: string
                example: "mystring"
        400:
          description: Bad Request
        500:
          description: Internal Server Error 

  /objects:
    get:
      tags:
        - Object
      summary: returns unique course info
      responses: 
        200:
          description: All good and success
          content:
            application/json:
              schema:
                type: object
                properties: 
                  id:
                    type: string
                  name:
                    type: string
                  price:
                    type: number
        400:
          description: Bad Request
        500:
          description: Internal Server Error 

  /courses:
    get:
      tags:
        - Array
      summary: returns all courses info
      responses: 
        200:
          description: All good and success
          content:
            application/json:
              schema:
                type: array
                items: 
                  type: object
                  properties: 
                    id:
                      type: string
                    name:
                      type: string
                    price:
                      type: number
        400:
          description: Bad Request
        500:
          description: Internal Server Error 

  /myCourse/{courseId}:
    get:
      tags:
        - String
      summary: returns course based on request id
      parameters: 
        - name: courseId
          in: path
          required: true
          default: 11
          schema:
            type: string
      responses: 
        200:
          description: All good and success
          content:
            application/json:
              schema: 
                type: object
                properties: 
                  id:
                    type: string
                  name:
                    type: string
                  price:
                    type: number
        400:
          description: Bad Request
        500:
          description: Internal Server Error 

  /addCourse:
    post:
      tags:
        - String
      summary: adds a new course to existing courses
      consumes:
        - application/json
      produces:
        - application/json
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties: 
                id:
                  type: string
                name:
                  type: string
                price:
                  type: number
      responses: 
        200:
          description: All good and success
          content:
            application/json:
              schema: 
                type: boolean
        400:
          description: Bad Request
        500:
          description: Internal Server Error 

  /courseQuery:
    get:
      tags:
        - String
      summary: trying to learn about query
      parameters:
        - name: location
          in: query
          required: true
          schema:
            type: string
            enum: [delhi, bangalore, mumbai]
        - name: device
          in: query
          required: true
          schema:
            type: string
            enum: [web, mobile]
      responses: 
        200:
          description: All good and success
          content:
            application/json:
              schema: 
                type: object
                properties: 
                  location:
                    type: string
                  device:
                    type: string
        400:
          description: Bad Request
        500:
          description: Internal Server Error 

  /courseUpload:
    post:
      tags:
        - String
      summary: Uploading course image
      parameters:
        - in: path
          name: auth
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file: 
                  type: string
                  format: binary
      responses: 
        200:
          description: All good and success
          content:
        400:
          description: Bad Request
        500:
          description: Internal Server Error 
                